userbot: Move init and "constants" variables into __init__.py
Easier to import the "constants" variables from other sub-modules
especially for `command` module.
